# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FitCycle Lifetime Fitness Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /exercises/{id}:
    get:
      operationId: getExercise
      tags: [exercises]
      summary: Get a specific exercise
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Exercise
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exercise"
  /workout-logs:
    get:
      operationId: getWorkoutLogs
      tags: [workout-logs]
      summary: Get workout logs with optional filters
      parameters:
        - name: startDate
          in: query
          schema:
            type: ["string", "null"]
        - name: endDate
          in: query
          schema:
            type: ["string", "null"]
        - name: dayNumber
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
        - name: offset
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Workout logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkoutLog"
    post:
      operationId: createWorkoutLog
      tags: [workout-logs]
      summary: Create a new workout log (finish workout)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkoutLogInput"
      responses:
        "201":
          description: Created workout log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutLog"
  /workout-logs/{id}:
    get:
      operationId: getWorkoutLog
      tags: [workout-logs]
      summary: Get a specific workout log with exercise logs
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Workout log detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutLogDetail"
    patch:
      operationId: updateWorkoutLog
      tags: [workout-logs]
      summary: Update a workout log
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkoutLogUpdate"
      responses:
        "200":
          description: Updated workout log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutLog"
    delete:
      operationId: deleteWorkoutLog
      tags: [workout-logs]
      summary: Delete a workout log
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /exercise-logs:
    post:
      operationId: createExerciseLog
      tags: [exercise-logs]
      summary: Create exercise log entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExerciseLogInput"
      responses:
        "201":
          description: Created exercise log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExerciseLog"
  /exercise-logs/{id}:
    patch:
      operationId: updateExerciseLog
      tags: [exercise-logs]
      summary: Update an exercise log entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExerciseLogUpdate"
      responses:
        "200":
          description: Updated exercise log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExerciseLog"
    delete:
      operationId: deleteExerciseLog
      tags: [exercise-logs]
      summary: Delete an exercise log
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /exercise-logs/by-workout/{workoutLogId}:
    get:
      operationId: getExerciseLogsByWorkout
      tags: [exercise-logs]
      summary: Get all exercise logs for a workout
      parameters:
        - name: workoutLogId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Exercise logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExerciseLog"
  /exercise-logs/progress:
    get:
      operationId: fetchExerciseHistory
      tags: [exercise-logs]
      summary: Get historical progress for an exercise
      parameters:
        - name: exerciseId
          in: query
          required: true
          schema:
            type: integer
        - name: days
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Exercise progress history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExerciseProgressSummary"
...
components:
  schemas:
...
    Exercise:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        trackingType:
          type: string
        defaultSets:
          type: ["integer", "null"]
        defaultReps:
          type: ["integer", "null"]
        defaultDurationSeconds:
          type: ["integer", "null"]
        defaultWeightKg:
          type: ["number", "null"]
        formTip:
          type: ["string", "null"]
      required: [id, name, category, trackingType]
...
    WorkoutLog:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
        cycleNumber:
          type: integer
        workoutDayNumber:
          type: integer
        status:
          type: string
        totalCalories:
          type: ["number", "null"]
        totalCyclingMinutes:
          type: ["number", "null"]
        energyLevel:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, date, cycleNumber, workoutDayNumber, status, createdAt]
    WorkoutLogDetail:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
        cycleNumber:
          type: integer
        workoutDayNumber:
          type: integer
        status:
          type: string
        totalCalories:
          type: ["number", "null"]
        totalCyclingMinutes:
          type: ["number", "null"]
        energyLevel:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        exerciseLogs:
          type: array
          items:
            $ref: "#/components/schemas/ExerciseLog"
      required: [id, date, cycleNumber, workoutDayNumber, status, createdAt, exerciseLogs]
    WorkoutLogInput:
      type: object
      properties:
        date:
          type: string
        cycleNumber:
          type: integer
        workoutDayNumber:
          type: integer
        status:
          type: string
        totalCalories:
          type: ["number", "null"]
        totalCyclingMinutes:
          type: ["number", "null"]
        energyLevel:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        exerciseLogs:
          type: array
          items:
            $ref: "#/components/schemas/ExerciseLogInput"
      required: [date, cycleNumber, workoutDayNumber, status]
    WorkoutLogUpdate:
      type: object
      properties:
        status:
          type: string
        totalCalories:
          type: ["number", "null"]
        totalCyclingMinutes:
          type: ["number", "null"]
        energyLevel:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    ExerciseLog:
      type: object
      properties:
        id:
          type: integer
        workoutLogId:
          type: integer
        exerciseId:
          type: integer
        status:
          type: string
        actualSets:
          type: ["integer", "null"]
        actualReps:
          type: ["integer", "null"]
        actualWeightKg:
          type: ["number", "null"]
        actualDurationSeconds:
          type: ["integer", "null"]
        totalVolume:
          type: ["number", "null"]
        formQuality:
          type: ["string", "null"]
        painDiscomfort:
          type: ["boolean", "null"]
        notes:
          type: ["string", "null"]
        exercise:
          $ref: "#/components/schemas/Exercise"
      required: [id, workoutLogId, exerciseId, status, exercise]
    ExerciseLogInput:
      type: object
      properties:
        workoutLogId:
          type: ["integer", "null"]
        exerciseId:
          type: integer
        status:
          type: string
        actualSets:
          type: ["integer", "null"]
        actualReps:
          type: ["integer", "null"]
        actualWeightKg:
          type: ["number", "null"]
        actualDurationSeconds:
          type: ["integer", "null"]
        totalVolume:
          type: ["number", "null"]
        formQuality:
          type: ["string", "null"]
        painDiscomfort:
          type: ["boolean", "null"]
        notes:
          type: ["string", "null"]
      required: [exerciseId, status]
    ExerciseLogUpdate:
      type: object
      properties:
        status:
          type: string
        actualSets:
          type: ["integer", "null"]
        actualReps:
          type: ["integer", "null"]
        actualWeightKg:
          type: ["number", "null"]
        actualDurationSeconds:
          type: ["integer", "null"]
        totalVolume:
          type: ["number", "null"]
        formQuality:
          type: ["string", "null"]
        painDiscomfort:
          type: ["boolean", "null"]
        notes:
          type: ["string", "null"]
    ExerciseProgressEntry:
      type: object
      properties:
        date:
          type: string
        actualSets:
          type: ["integer", "null"]
        actualReps:
          type: ["integer", "null"]
        actualWeightKg:
          type: ["number", "null"]
        actualDurationSeconds:
          type: ["integer", "null"]
        totalVolume:
          type: ["number", "null"]
        formQuality:
          type: ["string", "null"]
        cycleNumber:
          type: integer
      required: [date, cycleNumber]
    ExerciseProgressSummary:
      type: object
      properties:
        exerciseId:
          type: integer
        exerciseName:
          type: string
        history:
          type: array
          items:
            $ref: "#/components/schemas/ExerciseProgressEntry"
        personalBestWeight:
          type: ["number", "null"]
        personalBestReps:
          type: ["integer", "null"]
        personalBestVolume:
          type: ["number", "null"]
        lastPerformedDate:
          type: ["string", "null"]
        firstLoggedWeight:
          type: ["number", "null"]
        improvementPercent:
          type: ["number", "null"]
      required: [exerciseId, exerciseName, history]
...