# lib/api-zod/src/generated/types/exerciseLogUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FitCycle Lifetime Fitness Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface ExerciseLogUpdate {
  status?: string;
  /** @nullable */
  actualSets?: number | null;
  /** @nullable */
  actualReps?: number | null;
  /** @nullable */
  actualWeightKg?: number | null;
  /** @nullable */
  actualDurationSeconds?: number | null;
  /** @nullable */
  totalVolume?: number | null;
  /** @nullable */
  formQuality?: string | null;
  /** @nullable */
  painDiscomfort?: boolean | null;
  /** @nullable */
  notes?: string | null;
}