# lib/api-zod/src/generated/types/exerciseProgressSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FitCycle Lifetime Fitness Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { ExerciseProgressEntry } from './exerciseProgressEntry';

export interface ExerciseProgressSummary {
  exerciseId: number;
  exerciseName: string;
  history: ExerciseProgressEntry[];
  /** @nullable */
  personalBestWeight?: number | null;
  /** @nullable */
  personalBestReps?: number | null;
  /** @nullable */
  personalBestVolume?: number | null;
  /** @nullable */
  lastPerformedDate?: string | null;
  /** @nullable */
  firstLoggedWeight?: number | null;
  /** @nullable */
  improvementPercent?: number | null;
}